dockerfolder/requirements.txt
 fastapi==0.79.0
pickle==0.7.5
Pillow==9.1.0
torch==1.11.0
torchvision==0.12.0
transformers==4.20.1
uvicorn==0.18.2